Fix exception when creating model using get_model_by_id
get_model_by_id passes create=False which throws an undefined variable error on model_file_name since it doesn't exist in that branch.  Since the model is specified by ID, it must already exist and be uploaded, so the uploading feature can be skipped.  The change here allows it to work.